Move files around, refactor into packages
