+++ src/actiongraphgen/structure/node_data.py
